Added exception handling for player APIs

+++ cmpe275_lab2/src/main/java/edu/sjsu/cmpe275/lab2/dao/PlayerDaoImpl.java
@@ -1,6 +1,9 @@
 package edu.sjsu.cmpe275.lab2.dao;
 
+import edu.sjsu.cmpe275.lab2.exception.InvalidPlayerException;
+import edu.sjsu.cmpe275.lab2.exception.PlayerEmailInvalidException;
 import edu.sjsu.cmpe275.lab2.exception.PlayerNotFoundException;
+import edu.sjsu.cmpe275.lab2.exception.SponsorNotFoundException;
 import edu.sjsu.cmpe275.lab2.model.Player;
 import edu.sjsu.cmpe275.lab2.model.Sponsor;
 
@@ -23,24 +26,32 @@ public class PlayerDaoImpl implements PlayerDao {
   @Override
   public Player getPlayerById(Long id) {
     Player player = entityManager.find(Player.class, id);
-    if(player == null) throw new PlayerNotFoundException("Player not found");
+    if(player == null) throw new PlayerNotFoundException("Player Not Found");
     return player;
   }
   
   @Override
   public Player createPlayer(String firstname, String lastname, String email, String description, Long sponsorId) {
 	
+	if(firstname.isEmpty() || lastname.isEmpty() || email.isEmpty()) throw new InvalidPlayerException("Firstname, Lastname and Email cannot be empty");  
+	  
 	Player player = new Player();
 	player.setFirstname(firstname);
 	player.setLastname(lastname);
-	player.setEmail(email);
 	player.setDescription(description);
 	
 	if(sponsorId != null) {
 		Sponsor sponsor = entityManager.find(Sponsor.class, sponsorId);
+		if(sponsor == null) throw new SponsorNotFoundException("Sponsor Not Found");
 		player.setSponsor(sponsor);
-	}	
-
+	}
+	
+	Query query = entityManager.createQuery("from Player where email =: email")
+			.setParameter("email", email);	
+	Player playerWithSameEmail = (Player) query.getSingleResult();
+	if(playerWithSameEmail != null) throw new PlayerEmailInvalidException("Email ID is already taken");
+	
+	player.setEmail(email);
 	Player dbPlayer = entityManager.merge(player);
 
 	return dbPlayer;
@@ -48,18 +59,28 @@ public class PlayerDaoImpl implements PlayerDao {
   
   @Override
   public Player updatePlayer(Long id, String firstname, String lastname, String email, String description, Long sponsorId) {
-	Player player = getPlayerById(id);
+	if(firstname.isEmpty() || lastname.isEmpty() || email.isEmpty()) throw new InvalidPlayerException("Firstname, Lastname and Email cannot be empty");
+	
+	Player player = entityManager.find(Player.class, id);
+	if(player == null) throw new PlayerNotFoundException("Player Not Found");
+	
 	player.setFirstname(firstname);
 	player.setLastname(lastname);
-	player.setEmail(email);
 	if(description != null && !description.isEmpty()) {
 		player.setDescription(description);
 	}	
 	if(sponsorId != null) {
 		Sponsor sponsor = entityManager.find(Sponsor.class, sponsorId);
+		if(sponsor == null) throw new SponsorNotFoundException("Sponsor Not Found");
 		player.setSponsor(sponsor);
 	}
 	
+	Query query = entityManager.createQuery("from Player where email =: email")
+			.setParameter("email", email);
+	Player playerWithSameEmail = (Player) query.getSingleResult();
+	if(playerWithSameEmail != null && playerWithSameEmail.getId() != id) throw new PlayerEmailInvalidException("Email ID is already taken");
+	
+	player.setEmail(email);
 	Player dbPlayer = entityManager.merge(player);
 	
 	return dbPlayer;
@@ -67,6 +88,9 @@ public class PlayerDaoImpl implements PlayerDao {
   
   @Override
   public void deletePlayer(Long id) {
+	  
+	Player player = getPlayerById(id);
+	if(player == null) throw new PlayerNotFoundException("Player Not Found");
 	Query query = entityManager.createQuery("delete from Player where id =: id")
 			.setParameter("id", id);
 	query.executeUpdate();
